+++ src/main/java/br/edu/ufape/poo/academia/negocio/cadastro/CadastroInstrutor.java
@@ -5,7 +5,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import br.edu.ufape.poo.academia.dados.InstrutorRepository;
 import br.edu.ufape.poo.academia.negocio.basico.Instrutor;
-import br.edu.ufape.poo.academia.negocio.cadastro.InstrutorNaoEncontradoException;
 
 @Service
 public class CadastroInstrutor implements InterfaceCadastroInstrutor {
@@ -28,12 +27,21 @@ public class CadastroInstrutor implements InterfaceCadastroInstrutor {
     }
 
     @Override
+    public Instrutor procurarInstrutorPorId(Long id) throws InstrutorNaoEncontradoException {
+        return instrutorRepository.findById(id)
+                .orElseThrow(() -> new InstrutorNaoEncontradoException(id.toString()));
+    }
+
+    @Override
     public List<Instrutor> listarInstrutores() {
         return instrutorRepository.findAll();
     }
 
     @Override
-    public void removerInstrutor(Long id) {
+    public void removerInstrutor(Long id) throws InstrutorNaoEncontradoException {
+        if (!instrutorRepository.existsById(id)) {
+            throw new InstrutorNaoEncontradoException(id.toString());
+        }
         instrutorRepository.deleteById(id);
     }
 }
